requirement.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
Babel==2.3.4
backcall==0.1.0
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
decorator==4.0.10
docutils==0.12
ebaysdk==2.1.5
feedparser==5.2.1
greenlet==0.4.13
html2text==2016.9.19
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
Jinja2==2.8
lxml==4.2.1
Mako==1.0.4
MarkupSafe==0.23
mock==2.0.0
num2words==0.5.4
ofxparse==0.16
olefile==0.45.1
parsel==1.5.1
parso==0.2.1
passlib==1.6.5
pbr==4.0.4
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
py2exe==0.9.2.2
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyDispatcher==2.0.5
pydot==1.2.3
Pygments==2.2.0
PyMySQL==0.9.3
pyOpenSSL==18.0.0
pyparsing==2.1.10
PyPDF2==1.26.0
pypiwin32==223
pyserial==3.1.1
python-dateutil==2.5.3
python-stdnum==1.9
pytz==2016.7
pyusb==1.0.0
pywin32==223
PyYAML==3.12
qrcode==5.3
queuelib==1.5.0
reportlab==3.4.0
requests==2.11.1
Scrapy==1.5.1
selenium==3.141.0
service-identity==18.1.0
simplegeneric==0.8.1
six==1.11.0
suds-jurko==0.6
traitlets==4.3.2
Twisted==18.4.0
urllib3==1.23
vatnumber==1.2
vobject==0.9.3
w3lib==1.19.0
wcwidth==0.1.7
Werkzeug==0.11.15
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.3.0
zope.interface==4.5.0
